requirements.txt
fastapi
uvicorn
torch
diffusers
pillow
nest_asyncio
transformers
pydantic
accelerate
datasets